refactor: apply reviewed simplifications and edge-case fixes
Round 1 of an automated review+simplify pass over the branch diff.
Verified each change is behavior-preserving (or a confirmed edge-case
fix) and runs clean through `flutter analyze` + the test suite.

Bug/edge-case fixes:
- database_manager: track and await pending db close before deleteFor,
  closing an open/close race
- pull_sync: stop double-counting archived-with-synced-body chats in
  changedChats / cycle-done log
- media_upload_controller: clean up converted temp dir on the
  error/abort path instead of leaking it
- sync_triggers: seed foreground state on cold launch so the periodic
  foreground pull starts when launched already foregrounded

Simplifications / dead-code removal:
- auth_state_manager: drop unused params and duplicate error log
- notes_dao: remove dead `meta` param from updateNoteWithOutbox
- search_dao: drop unused `kind` projection from the hits CTE
- fts_query: make _quoteToken non-nullable, remove unreachable guard
- note_mapper / note_sync: collapse duplicate _decodeMap/_asNs helpers
  onto the shared decodeJsonMap/asNs
- outbox_task_queue_migrator: fetch chat row once instead of twice
- push_sync / sync_api_client: remove always-true guard and unreachable
  404 branch
- backoff: drop redundant .toInt()
- chat_transport_dispatch / request_completion_runner /
  local_conversation_loader: simplify return, fix stale doc comments,
  drop redundant cast/import
- chat_page: fire attachment uploads non-blocking with error logging

+++ lib/core/sync/backoff.dart
@@ -35,7 +35,7 @@ class Backoff {
   int delayMsForAttempt(int attempt) {
     // Clamp the shift to avoid 1<<n overflow; the window saturates at capMs
     // long before attempt reaches 30 anyway.
-    final shift = attempt.clamp(0, 30).toInt();
+    final shift = attempt.clamp(0, 30);
     final window = (baseMs * (1 << shift)).clamp(baseMs, capMs);
     final sample = jitter();
     final boundedJitter = sample.isNaN
